move all composing views into AppFactory
